Widget Ownership is now checked when editing / deleting widgets!
New function 'DoIOwnThisWidget' added under shared functions

+++ NewWidget.php
@@ -18,52 +18,64 @@
     }
 
     $post_data = file_get_contents('php://input');
+    $break_if_tried_editing_widget_without_permission = false;
 
 	//Delete Existing Widget ID if Edit ID passed: 
 	If ($_POST["ID"] <> "") {
-		$sql = "Delete From Widgets Where RecID = '" . $_POST["ID"] . "'";
-		execquery($sql);
-		echo "DELETED WIDGET WITH ID '" . $_POST["ID"] . "', Inserting new copy now...";
+        $widgetID = $_POST["ID"];
+        if (DoIOwnThisWidget($widgetID)) {
+            $sql = "Delete From Widgets Where RecID = '" . $_POST["ID"] . "'";
+            execquery($sql);
+            echo "DELETED WIDGET WITH ID '" . $_POST["ID"] . "', Inserting new copy now...";
+        } else {
+            $break_if_tried_editing_widget_without_permission = true;
+            header("Location: index.php");
+        }
+		
 		header("Location: index.php");
 	}
+    if (!$break_if_tried_editing_widget_without_permission) {
+        //Insert new Widget
+        $dashboardid = $_POST["dashboardID"]; //Identify which dashboard the widget needs to be added to.
 
-    //Insert new Widget
-	$dashboardid = $_POST["dashboardID"]; //Identify which dashboard the widget needs to be added to.
+        //Prepare Variables
+        $sep = "','"; // Seperator
+        $sqlserveraddress = $_POST["SQLServerAddressName"];
+        $sqldbname = $_POST["SQLDBName"];
+        $sqlusername = $_POST["sqluser"];
+        $sqlpass = $_POST["sqlpass"];
+        $sqlquery = str_replace("'","''",$_POST["sqlquery"]);
+        
+        $globaldefault = "";
 
-    //Prepare Variables
-    $sep = "','"; // Seperator
-    $sqlserveraddress = $_POST["SQLServerAddressName"];
-    $sqldbname = $_POST["SQLDBName"];
-    $sqlusername = $_POST["sqluser"];
-    $sqlpass = $_POST["sqlpass"];
-    $sqlquery = str_replace("'","''",$_POST["sqlquery"]);
-    
-    $globaldefault = "";
+        if (isset($_POST["GlobalDefault"])) {
+            $globaldefault = $_POST["GlobalDefault"];
 
-    if (isset($_POST["GlobalDefault"])) {
-        $globaldefault = $_POST["GlobalDefault"];
+        } else {
+            $globaldefault = "0";
+        }
 
+        //$_POST[""]
+        
+        // Prepare INSERT statement.
+        $select = "INSERT INTO Widgets (WidgetType,BookmarkDisplayText,PositionX,PositionY,SizeX,SizeY,WidgetURL,WidgetCSSClass,Notes,DashboardRecID
+        ,sqlserveraddress,sqldbname,sqluser,sqlpass,sqlquery, Global) 
+        VALUES ('" . $_POST["WidgetType"] . $sep . $_POST["DisplayText"] . $sep . $_POST["PositionX"] . $sep . $_POST["PositionY"] . $sep . 
+        $_POST["SizeX"] . $sep . $_POST["SizeY"] . $sep . $_POST["URL"] . $sep . $_POST["CSSClass"] . $sep . str_replace("'", "''",$_POST["Notes"]) . $sep . $dashboardid . $sep .
+        $sqlserveraddress . $sep . $sqldbname . $sep . $sqlusername . $sep . $sqlpass . $sep . $sqlquery .
+        "', '" . $globaldefault . "')";
+        debuglog($select);
+        debuglog($sessionID, "Session ID");
+        debuglog($dashboardid, "Dashboard ID");
+
+        //$stmt->bindParam(':notes', $_POST["Notes"]); // Failed, don't understand why.
+        execquery($select);
+            //echo "Finished; click <a href='index.php'>Here</a> to return to main dashboard.";
+	    header("Location: index.php" . "?SelectDashboardID=" . $dashboardid);
     } else {
-        $globaldefault = "0";
+        header("Location: index.php");
     }
 
-    //$_POST[""]
-	
-    // Prepare INSERT statement.
-	$select = "INSERT INTO Widgets (WidgetType,BookmarkDisplayText,PositionX,PositionY,SizeX,SizeY,WidgetURL,WidgetCSSClass,Notes,DashboardRecID
-    ,sqlserveraddress,sqldbname,sqluser,sqlpass,sqlquery, Global) 
-    VALUES ('" . $_POST["WidgetType"] . $sep . $_POST["DisplayText"] . $sep . $_POST["PositionX"] . $sep . $_POST["PositionY"] . $sep . 
-    $_POST["SizeX"] . $sep . $_POST["SizeY"] . $sep . $_POST["URL"] . $sep . $_POST["CSSClass"] . $sep . str_replace("'", "''",$_POST["Notes"]) . $sep . $dashboardid . $sep .
-    $sqlserveraddress . $sep . $sqldbname . $sep . $sqlusername . $sep . $sqlpass . $sep . $sqlquery .
-    "', '" . $globaldefault . "')";
-	debuglog($select);
-    debuglog($sessionID, "Session ID");
-    debuglog($dashboardid, "Dashboard ID");
-
-    //$stmt->bindParam(':notes', $_POST["Notes"]); // Failed, don't understand why.
-    execquery($select);
-    //echo "Finished; click <a href='index.php'>Here</a> to return to main dashboard.";
-	header("Location: index.php" . "?SelectDashboardID=" . $dashboardid);
 ?>
 
 </body>
