+++ actions/DeleteWidget.php
@@ -6,10 +6,17 @@
 	$qr = $_SERVER['QUERY_STRING']; // Get querystring value
 	$d = str_replace("RecID=","",$qr); // strip out querystring key / header, just want the value
     
-	$select = "Delete From Widgets Where RecID = '" . $d . "'";
-	execquery($select);
+	if (DoIOwnThisWidget($d)) {
+
+		$select = "Delete From Widgets Where RecID = '" . $d . "'";
+		execquery($select);
+		
+
+	} else {
+
+	}
 	echo "<script>window.close();</script>Complete. Window should close now. <br />Query Executed: " . $select;
-	
+		
 	echo "<a href='http://douglasrobinson.me/Dashboardify/'>Return to Management</a>";
 	header("Location: ../index.php");
 ?>
